+++ src/components/CampaignTimeAndSalary/CampaignTimeAndSalaryBoard/index.js
@@ -38,6 +38,8 @@ import {
 } from '../../TimeAndSalary/TimeAndSalaryBoard/helper';
 import { DATA_NUM_PER_PAGE } from '../../../constants/timeAndSalarSearch';
 
+import renderHelmet from './helmet';
+
 const pathnameMapping = {
   '/time-and-salary/campaigns/:campaign_name/work-time-dashboard': {
     title: '工時排行榜',
@@ -284,6 +286,8 @@ export default class CampaignTimeAndSalaryBoard extends Component {
     } = this.props;
     const { title } = pathnameMapping[path];
     const { status, data, switchPath, totalCount, currentPage } = this.props;
+    const pathname = this.props.location.pathname;
+    const campaignInfo = campaignEntries.get(campaignName).toJS();
     const raw = data.toJS();
 
     // 如果 campaignName 不在清單中，代表 Not Found
@@ -300,6 +304,7 @@ export default class CampaignTimeAndSalaryBoard extends Component {
 
     return (
       <section className={timeAndSalaryCommonStyles.searchResult}>
+        {renderHelmet({ pathname, title, campaignInfo })}
         <h2 className={styles.heading}>{title}</h2>
         <Link
           className={cn(
